% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.svi.R
\docType{data}
\name{ca.svi}
\alias{ca.svi}
\title{2018 CDC Social Vulnerability Index for California}
\format{
\subsection{\code{ca.svi}}{

A data frame with 8,041 rows and 124 columns.
\describe{
\item{STATE}{state}
\item{FIPS}{state and county FIPS code.}
...
}
}
}
\source{
\url{https://www.atsdr.cdc.gov/placeandhealth/svi/index.html}
}
\usage{
ca.svi
}
\description{
2018 CDC Social Vulnerability Index for California
}
\keyword{datasets}
