% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\docType{data}
\name{conditions}
\alias{conditions}
\title{Conditions table from ctrialsgov duckdb with 2 columns dropped}
\format{
\subsection{\code{conditions}}{

A data frame with 815,944 rows and 2 columns:
\describe{
\item{nct_id}{Unique identifier assigned to each study on clinicaltrials.gov}
\item{name}{Name of condition}
}
}
}
\usage{
conditions
}
\description{
Conditions table from ctrialsgov duckdb with 2 columns dropped
}
\keyword{datasets}
