% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\docType{data}
\name{countries}
\alias{countries}
\title{Countries table from ctrialsgov duckdb with 1 column dropped}
\format{
\subsection{\code{countries}}{

A data frame with 653,881 rows and 3 columns:
\describe{
\item{nct_id}{Unique identifier assigned to each study on clinicaltrials.gov}
\item{name}{Name of country}
\item{removed}{Whether country was removed or not from the study}
}
}
}
\usage{
countries
}
\description{
Countries table from ctrialsgov duckdb with 1 column dropped
}
\keyword{datasets}
