% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_conditions_histogram.R
\name{create_conditions_histogram}
\alias{create_conditions_histogram}
\title{Plots histogram of conditions examined in the studies}
\usage{
create_conditions_histogram(d, conditions)
}
\arguments{
\item{d}{the dataframe of the query result}

\item{conditions}{the handle referencing the "conditions" database table}
}
\value{
ggplot object that shows histogram of the examined conditions distribution
}
\description{
Displays a histogram to show the distribution of conditions examined in the
studies in a query
}
