% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_endpoint_histogram.R
\name{create_endpoint_histogram}
\alias{create_endpoint_histogram}
\title{Plots histogram of endpoint status in the studies}
\usage{
create_endpoint_histogram(d, endpoints)
}
\arguments{
\item{d}{the dataframe of the query result}

\item{endpoints}{the dataframe that contains endpoint information of the studies}
}
\value{
ggplot object that shows histogram of the distribution of whether the
endpoint was met or not
}
\description{
Displays a histogram to show the distribution of whether the endpoint was
met or not in the studies in a query
}
