% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phase_histogram.R
\name{create_phase_histogram}
\alias{create_phase_histogram}
\title{Plots histogram of phases of the studies}
\usage{
create_phase_histogram(d, studies)
}
\arguments{
\item{d}{the dataframe of the query result}

\item{studies}{the handle referencing the "studies" database table}
}
\value{
ggplot object that shows histogram of trial phase distribution
}
\description{
Displays a histogram to show the distribution of phases from the studies in a query
}
