% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_studytype_histogram.R
\name{create_study_type_histogram}
\alias{create_study_type_histogram}
\title{Plots histogram of study types of the studies}
\usage{
create_study_type_histogram(d)
}
\arguments{
\item{d}{the dataframe of the query result}
}
\value{
ggplot object that shows histogram of the study type distribution
}
\description{
Displays a histogram to show the distribution of study types of the
studies in a query
}
