% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topcountries_table.R
\name{create_topcountries_table}
\alias{create_topcountries_table}
\title{Outputs the most countries involved in studies in descending order}
\usage{
create_topcountries_table(d, countries)
}
\arguments{
\item{d}{the dataframe of the query result}

\item{countries}{the handle referencing the "countries" database table}
}
\value{
dataframe object that lists the countries involved in the studies
in descending order of involvement
}
\description{
Displays a table of countries in decreasing order of involvement (weren't
identified as having been removed) from the studies in a query
}
