% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\docType{data}
\name{endpoints}
\alias{endpoints}
\title{Endpoints of studies from ctrialsgov duckdb}
\format{
\subsection{\code{endpoints}}{

A data frame with 6,084 rows and 3 columns:
\describe{
\item{nct_id}{Unique identifier assigned to each study on clinicaltrials.gov}
\item{endpoint_met}{Whether endpoint was met or not}
}
}
}
\usage{
endpoints
}
\description{
Endpoints of studies from ctrialsgov duckdb
}
\keyword{datasets}
