% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Run Shiny app}
\usage{
launchApp()
}
\arguments{
\item{con}{the connection to the ctrialsgov duckdb file}
}
\description{
Starts Shiny app that allows user to query ctrialsgov database and see
relevant histograms and tables
}
