% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\docType{data}
\name{studies}
\alias{studies}
\title{Subset of studies table from ctrialsgov duckdb}
\format{
\subsection{\code{studies}}{

A data frame with 25,000 rows and 70 columns:
\describe{
\item{nct_id}{Unique identifier assigned to each study on clinicaltrials.gov}
\item{endpoint_met}{Whether endpoint was met or not}
}
}
}
\usage{
studies
}
\description{
Subset of studies table from ctrialsgov duckdb
}
\keyword{datasets}
