% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\docType{data}
\name{risk}
\alias{risk}
\title{Wildfire Hazard Risk Data}
\format{
\subsection{\code{risk}}{

A data frame with 3,108 rows and 7 columns.
\describe{
\item{GeoType}{All counties in this dataset}
\item{StateFP}{State FIPS code}
\item{GEOID}{State and county FIPS code}
\item{County_name}{County name}
\item{TotalPop}{Total population in county}
\item{state}{State name}
\item{cc_fire_rating}{Wildfire risk}
}
}
}
\usage{
risk
}
\description{
Wildfire Hazard Risk Data
}
\keyword{datasets}
